Move GraphQLTypeAsync for Arc<T> into pointers.rs
